view: fix topic threading when ghosts are present
This fixes a bug where a message replying to a ghost would
accidentally be added to the wrong topic in the index/topic
view.

Before commit 76d8f68dc273e54809ad69cfe49e141003f790ef ("view:
avoid recursion in topic index"), we would refuse to indent a
topic which started with a ghost which hid the problem.  Now we
inform the user the thread started elsewhere.