filter/rubylang: do not set altid on spam training
I suppose it's a bug or inconsistency that altid is write-only
and their deletions do not get reflected.  But for now, we
do not set it when training spam so there's no window where
an invalid NNTP article number shows up.

This should solve the problem where there's massive gaps
in messages solved by spam training for ruby groups:

	https://public-inbox.org/meta/20180307093754.GA27748@dcvr/