convert: copy description and git config from v1 repo
I noticed I lost a $GIT_DIR/description in a conversion, so we
should preserve it.  While we're at it, we ought to copy any
config in the old repo to the new one.

We will need to warn about cloneurl since it's unfortunately
not an automatic process to update.  Oh well..