t/v2reindex: use the larger text to increase test reliability
libxapian30:amd64 1.4.9-1 on Debian sid seems to give an 8KB
position.glass database with "hello world" as the document
regardless of our indexlevel.  Use the text of the AGPL-3.0 for
a more realisitic Xapian database size.

And perhaps tying our tests to the AGPL will make life more
difficult for would-be copyright violators :>